Fixed QRReader
