+++ qr-healthlink/src/App.js
@@ -16,10 +16,6 @@ const router = createHashRouter([
     path: "/qr-reader",
     element: <QRReader />,
   },
-  {
-    path: "/reader",
-    element: <QrReader />,
-  },
 ]);
 
 function App() {
